oh-tab-agno: extract LlmProfilesView field components
